Fix already existing folders causing issues

+++ tools/build_conda.py
@@ -15,6 +15,10 @@ class FileMover():
         self.destination_root = destination_root
 
     def move_file(self, src, dst):
+        if os.path.isdir(src) and os.path.exists(dst) and os.path.isdir(dst):
+            os.write(
+                1, "(skipped - already exists) move {} {}\n".format(src, dst).encode())
+            return
         os.write(1, "move {} {}\n".format(src, dst).encode())
         shutil.move(src, dst)
 
@@ -25,9 +29,9 @@ class FileMover():
             dst = glob.glob(dst)[-1]
         if '*' in src:
             for f in glob.glob(src):
-                self.move_file(f, dst)
+                self.move_file(f, os.path.join(dst, os.path.split(f)[1]))
         else:
-            self.move_file(src, dst)
+            self.move_file(src, os.path.join(dst, os.path.split(src)[1]))
 
 
 if __name__ == '__main__':
@@ -49,10 +53,10 @@ if __name__ == '__main__':
 
     # Depending on the platform, directories have different names.
     if sys.platform == "win32":
-        lib_dest = 'lib'
+        lib_dest = 'lib\\'
         dll_src = os.path.join("Lib", "scipp")
         dll_dest = os.path.join("Lib", "scipp")
-        lib_src = 'Lib'
+        lib_src = 'Lib\\'
         inc_src = 'include'
     else:
         lib_dest = os.path.join('lib', 'python*')
